some changes in the url structure and folders.
