+++ acbc_app/academia_blockchain/api/api.py
@@ -10,4 +10,6 @@ class ApiRoot(APIView):
         return Response({
             'profiles': request.build_absolute_uri(reverse_lazy('profile-list')),
             'courses': request.build_absolute_uri(reverse_lazy('course-list')),
+            'content': request.build_absolute_uri(reverse_lazy('library-list')),
+            'group': request.build_absolute_uri(reverse_lazy('group-list')),
         })
