+++ acbc_app/content/api/views.py
@@ -1,0 +1,117 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.generics import get_object_or_404
+from content.models import Library, File
+from .serializers import LibrarySerializer, FileSerializer
+
+
+class LibraryList(APIView):
+    def get(self, request, format=None):
+        libraries = Library.objects.all()
+        serializer = LibrarySerializer(libraries, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, format=None):
+        serializer = LibrarySerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class LibraryDetail(APIView):
+    def get_object(self, pk):
+        return get_object_or_404(Library, pk=pk)
+
+    def get(self, request, pk, format=None):
+        library = self.get_object(pk)
+        serializer = LibrarySerializer(library)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        library = self.get_object(pk)
+        serializer = LibrarySerializer(library, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        library = self.get_object(pk)
+        library.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+from content.models import Group
+from .serializers import GroupSerializer
+
+class GroupList(APIView):
+    def get(self, request, format=None):
+        groups = Group.objects.all()
+        serializer = GroupSerializer(groups, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, format=None):
+        serializer = GroupSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class GroupDetail(APIView):
+    def get_object(self, pk):
+        return get_object_or_404(Group, pk=pk)
+
+    def get(self, request, pk, format=None):
+        group = self.get_object(pk)
+        serializer = GroupSerializer(group)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        group = self.get_object(pk)
+        serializer = GroupSerializer(group, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        group = self.get_object(pk)
+        group.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class FileList(APIView):
+    def get(self, request, format=None):
+        files = File.objects.all()
+        serializer = FileSerializer(files, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, format=None):
+        serializer = FileSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class FileDetail(APIView):
+    def get_object(self, pk):
+        return get_object_or_404(File, pk=pk)
+
+    def get(self, request, pk, format=None):
+        file = self.get_object(pk)
+        serializer = FileSerializer(file)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        file = self.get_object(pk)
+        serializer = FileSerializer(file, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        file = self.get_object(pk)
+        file.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
