--- acbc_app/content/api_urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-from content.api_views import (LibraryList, LibraryDetail, GroupList, GroupDetail,
-                                    FileList, FileDetail)
-
-urlpatterns = [
-    path('libraries/', LibraryList.as_view(), name='library-list'),
-    path('libraries/<int:pk>/', LibraryDetail.as_view(), name='library-detail'),
-    path('groups/', GroupList.as_view(), name='group-list'),
-    path('groups/<int:pk>/', GroupDetail.as_view(), name='group-detail'),
-    path('files/', FileList.as_view(), name='file-list'),
-    path('files/<int:pk>/', FileDetail.as_view(), name='file-detail'),
-]
