+++ acbc_app/profiles/api/views.py
@@ -1,13 +1,25 @@
-from rest_framework.viewsets import ModelViewSet
-from ..models import Profile, AcceptedCrypto
-from .serializers import ProfileSerializer, AcceptedCryptoSerializer
+# profiles/api/views.py
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import status
+from profiles.models import Profile
+from .serializers import ProfileSerializer
+from django.shortcuts import get_object_or_404
 
+class ProfileList(APIView):
+    def get(self, request, format=None):
+        profiles = Profile.objects.all()
+        serializer = ProfileSerializer(profiles, many=True)
+        return Response(serializer.data)
 
-class ProfileList(ModelViewSet):
-    queryset = Profile.objects.all()
-    serializer_class = ProfileSerializer
+    def post(self, request, format=None):
+        serializer = ProfileSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class ProfileDetail(ModelViewSet):
+class ProfileDetail(APIView):
     def get(self, request, pk, format=None):
         profile = get_object_or_404(Profile, pk=pk)
         serializer = ProfileSerializer(profile)
@@ -21,22 +33,16 @@ class ProfileDetail(ModelViewSet):
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    def patch(self, request, pk, format=None):
+        profile = get_object_or_404(Profile, pk=pk)
+        serializer = ProfileSerializer(profile, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
     def delete(self, request, pk, format=None):
         profile = get_object_or_404(Profile, pk=pk)
         profile.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class AcceptedCryptoList(ModelViewSet):
-    def get(self, request, format=None):
-        cryptos = AcceptedCrypto.objects.all()
-        serializer = AcceptedCryptoSerializer(cryptos, many=True)
-        return Response(serializer.data)
-
-    def post(self, request, format=None):
-        serializer = AcceptedCryptoSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
