+++ frontend/src/api/axios.ts
@@ -1,7 +1,7 @@
 // src/api/axios.ts
 import axios from 'axios';
 
-const baseURL = 'http://localhost:8000/';
+const baseURL = 'http://localhost:8000/api';
 
 const axiosInstance = axios.create({
   baseURL: baseURL,
