--- frontend/src/config/axios.tsx
@@ -1,8 +1,0 @@
-import axios from 'axios'
-
-
-const clienteAxios= axios.create({
-    baseURL:'http://127.0.0.1:8000/api',
-});
-
-export default clienteAxios;
