--- frontend/src/paginas/events.tsx
@@ -1,35 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import clienteAxios from '../config/axios';
-import { useNavigate } from 'react-router-dom';
-
-
-const Events = () => {
-
-
-
-
-
-  return (
-    <div className="courses-events-container">
-      <div className="text-center mt-6">
-        <button
-          type="button"
-          className="btn btn-warning text-uppercase"
-
-        >
-          Create New Course or Event
-        </button>
-      </div>
-
-      <p>
-        <strong>Your Courses and Events:</strong>
-        <br />
-      </p>
-
-      {/* Aquí puedes agregar más contenido para mostrar los cursos y eventos */}
-    </div>
-  );
-};
-
-
-export default Events
