--- frontend/src/paginas/security.tsx
@@ -1,19 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import clienteAxios from '../config/axios';
-import { useNavigate } from 'react-router-dom';
-
-const Security = () => {
-
-
-    console.log('hola mundo')
-
-    return (
-        <div>
-                <p> Información del perfil </p>
-        </div>
-
-
-        );
-};
-
-export default Security;
